+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file.
 `APExtensions` adheres to [Semantic Versioning](http://semver.org/).
 
+## [10.1.7](https://github.com/APUtils/APExtensions/releases/tag/10.1.7)
+Released on 18.11.2021.
+
+#### Added
+- NSAttributedString: Occupiable
+- NSSet: Occupiable
+- NSString: Occupiable
+
+
 ## [10.1.5](https://github.com/APUtils/APExtensions/releases/tag/10.1.5)
 Released on 17.11.2021.